// src/contract/type.ts
export interface IGetBalanceParams {
  symbol: string;
  owner: string;
}

export interface IGetTokenInfoParams {
  symbol: string;
}
// ...
export enum ContractMethodType {
  SEND,
  VIEW,
}

export interface IContractOptions {
  chain?: Chain | null;
  type?: ContractMethodType;
}

export interface IGetAllowanceResponse {
  symbol: string;
  owner: string;
  spender: string;
  allowance: number;
}

export interface IContractError extends Error {
  code?: number;
  error?:
    | number
    | string
    | {
        message?: string;
      };
  errorMessage?: {
    message: string;
    name?: string;
    stack?: string;
  };
  Error?: string;
  from?: string;
  sid?: string;
  result?: {
    TransactionId: string;
    transactionId?: string;
  };
  TransactionId?: string;
  transactionId?: string;
  value?: any;
}

export interface IApproveParams {
  spender: string;
  symbol: string;
  amount: number | string;
  batchApproveNFT?: boolean;
}
// ...
export interface ITransactionLog {
  Address: string;
  Name: string;
  Indexed: string[];
  NonIndexed: string;
}

export interface ITransactionResult {
  TransactionId: string;
  Status: string;
  Logs: ITransactionLog[];
  [props: string]: any;
}

export interface ISendResult {
  TransactionId: string;
  TransactionResult: ITransactionResult;
}